requirements.txt
attr==0.3.2
ConfigParser==7.2.0
contextlib2==21.6.0
cryptography==46.0.5
Cython==3.2.4
dl==0.1.0
docutils==0.22.4
h5py==3.14.0
HTMLParser==0.0.2
ipython==8.12.3
ipywidgets==8.1.8
Jinja2==3.1.6
jnius==1.1.0
keyring==25.7.0
lockfile==0.12.2
matplotlib==3.10.8
mock==5.2.0
numpy==1.24.3
opencv_python==4.10.0.84
Pillow==12.1.1
protobuf==7.34.0
psutil==7.2.2
pyOpenSSL==25.3.0
pywin32==311
railroad==0.5.0
referencing==0.37.0
scikit_learn==1.8.0
scipy==1.17.1
Sphinx==9.1.0
thread==2.0.6
torch==2.7.1+cu118
torchvision==0.22.1+cu118
tornado==6.5.1
trove_classifiers==2026.1.14.14
truststore==0.10.4
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
